Split occ_faceter into library and executable
